results/db: move from server/results to the top level
